refactor(fe/basic): split lowering helpers
